Kassenbuch Bugfixes

+++ DeleteBuchungsart.php
@@ -1,10 +1,9 @@
 <head>
-  <title>Kassenbuch Buchungsarten - Eintrag löschen</title>
+  <title>Kassenbuch Buchungsarten - Position löschen</title>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
   <!-- CSS -->
   <link href="css/bootstrap.min.css" rel="stylesheet">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
 
   <!-- JS -->
   <script src="js/jquery.min.js"></script>
@@ -12,22 +11,23 @@
 </head>
 
 <body>
+  
   <?php
   require 'db.php';
   session_start();
-
-  if (isset($_POST['id'])) {
+  
+  if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['id'])) {
     $id = intval($_POST['id']);
     $sql = "DELETE FROM buchungsarten WHERE id = :id";
     $stmt = $pdo->prepare($sql);
-    $stmt-"Kassenbuch Buchungsarten - Eintrag mit der ID" . $id . " wurde gelöscht!";
+    $stmt->execute(['id' => $id]);    
     sleep(1);
-    header('Location: Buchungsarten.php'); // Zurück zur Übersicht
-  
-    exit();
+    header('Location: Buchungsarten.php'); 
+    
   } else {
     echo "Ungültige Anfrage.";
   }
+
 
   ?>
 </body>
